+++ data/resources/CHANGELOG.md
@@ -1,6 +1,13 @@
+Jan 12, 2025:
+- test: Update for patch
+- Added ${Pet.ProcHold} and ${Pet.Resume}
+
 Jan 11, 2025:
 - ImGui has been updated to version 1.91.6
 - Fix date format string of output from /mqlog
+
+Dec 13, 2024:
+- test: Update for test patch
 
 Dec 5, 2024:
 - live: Update for patch